.cmdcd now works with customreactions too

+++ WizBot.Core/Services/Database/Repositories/Impl/PollsRepository.cs
@@ -24,16 +24,22 @@ namespace WizBot.Core.Services.Database.Repositories.Impl
                 .Include(x => x.Answers)
                 .Include(x => x.Votes)
                 .FirstOrDefault(x => x.Id == id);
+            
+            if (p is null)
+                return;
+
             if (p.Votes != null)
             {
                 _context.Set<PollVote>().RemoveRange(p.Votes);
                 p.Votes.Clear();
             }
+            
             if (p.Answers != null)
             {
                 _context.Set<PollAnswer>().RemoveRange(p.Answers);
                 p.Answers.Clear();
             }
+            
             _set.Remove(p);
         }
     }
